# public/openapi-gpt.json
{
  "openapi": "3.1.0",
  "info": { "title": "PaperflowAI Offer API", "version": "1.0.2" },
  "servers": [{ "url": "https://paperflowai-frontend.vercel.app", "description": "production" }],
  "paths": {
    "/api/offers/create": {
      "post": {
        "operationId": "createOffer",
        "summary": "Create offer and generate PDF",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customerId"],
                "properties": {
                  "customerId": { "type": "string", "description": "Kundkort-ID i PaperflowAI" },
                  "title": { "type": "string" },
                  "amount": { "type": "number" },
                  "currency": { "type": "string", "default": "SEK" },
                  "needsPrint": { "type": "boolean", "default": false },
                  "dataJson": { "type": "string", "description": "Stringified JSON with full offer content" }
                }
              },
              "example": {
                "customerId": "test-customer-123",
                "title": "Offert – Takjobb",
                "currency": "SEK",
                "needsPrint": false,
                "dataJson": "{\"rader\":[{\"timmar\":10,\"pris_per_timme\":900},{\"timmar\":50,\"pris_per_timme\":900}]}"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Offert skapad",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok", "offer"],
                  "properties": {
                    "ok": { "type": "boolean" },
                    "offer": {
                      "type": "object",
                      "required": ["id", "customer_id", "file_url", "created_at"],
                      "properties": {
                        "id": { "type": "string" },
                        "customer_id": { "type": "string" },
                        "title": { "type": "string" },
                        "amount": { "type": "number" },
                        "currency": { "type": "string" },
                        "file_url": { "type": "string", "format": "uri" },
                        "customer_url": { "type": "string", "format": "uri", "description": "Direktlänk till kundkortet" },
                        "created_at": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}